Consolidate all test only mutable manifest methods to utils
